test: enable ViewCommandHandlers-integ-test
